Enum quill_parser::expr_pat::ExprPatP

pub enum ExprPatP {
    Variable(IdentifierP),
    Immediate {
        range: Range,
        value: ConstantValue,
    },
    Apply(Box<ExprPatP>, Box<ExprPatP>),
    Lambda {
        lambda_token: Range,
        params: Vec<NameP>,
        expr: Box<ExprPatP>,
    },
    Let {
        let_token: Range,
        name: NameP,
        expr: Box<ExprPatP>,
    },
    Block {
        open_bracket: Range,
        close_bracket: Range,
        statements: Vec<ExprPatP>,
    },
    Borrow {
        borrow_token: Range,
        expr: Box<ExprPatP>,
    },
    Copy {
        copy_token: Range,
        expr: Box<ExprPatP>,
    },
    ConstructData {
        data_constructor: IdentifierP,
        open_brace: Range,
        close_brace: Range,
        fields: ConstructDataFields,
    },
    Impl {
        impl_token: Range,
        body: DefinitionBodyP,
    },
    ImplPattern {
        impl_token: Range,
        open_brace: Range,
        close_brace: Range,
        fields: ConstructDataFields,
    },
    Unknown(Range),
}

Represents either an expression or a pattern.

Variants

Variable(IdentifierP)

A named variable e.g. x or +.

Immediate

A primitive constant such as 14 or false.

Fields of Immediate

range: Rangevalue: ConstantValue

Apply(Box<ExprPatP>, Box<ExprPatP>)

Apply the left hand side to the right hand side, e.g. a b. More complicated expressions e.g. a b c d can be desugared into ((a b) c) d.

Lambda

A lambda abstraction, specifically lambda params -> expr.

Fields of Lambda

lambda_token: Rangeparams: Vec<NameP>expr: Box<ExprPatP>

Let

A let statement, specifically let identifier = left_expr;. This kind of statement is only valid as an intermediate statement in a block.

Fields of Let

let_token: Rangename: NamePexpr: Box<ExprPatP>

Block

A block of statements, inside parentheses, separated by newlines, where the final expression in the block is the type of the block as a whole.

Fields of Block

open_bracket: Rangeclose_bracket: Rangestatements: Vec<ExprPatP>

Borrow

Borrow some data for a duration less than its full lifetime.

Fields of Borrow

borrow_token: Rangeexpr: Box<ExprPatP>

Copy

Copy some data behind a borrow.

Fields of Copy

copy_token: Rangeexpr: Box<ExprPatP>

ConstructData

The name of a data type, followed by brace brackets containing the data structure’s fields.

Fields of ConstructData

data_constructor: IdentifierPopen_brace: Rangeclose_brace: Rangefields: ConstructDataFields

Impl

An implementation of an aspect. Only used in expressions, not patterns (see ExprPatP::ImplPattern).

Fields of Impl

impl_token: Rangebody: DefinitionBodyP

ImplPattern

A pattern destructuring an implementation of an aspect. Only used in patterns, not expressions (see ExprPatP::Impl).

Fields of ImplPattern

impl_token: Rangeopen_brace: Rangeclose_brace: Rangefields: ConstructDataFields

Unknown(Range)

An underscore _ representing an unknown. This is valid only in patterns, not normal expressions.