Enum quill_parser::expr_pat::ExprPatP

pub enum ExprPatP {

    Variable(IdentifierP),
    Constant {
        range: Range,
        value: ConstantValue,
    },
    String {
        range: Range,
        value: String,
    },
    Apply(Box<ExprPatP>, Box<ExprPatP>),
    Lambda {
        lambda_token: Range,
        params: Vec<NameP>,
        expr: Box<ExprPatP>,
    },
    Let {
        let_token: Range,
        name: NameP,
        expr: Box<ExprPatP>,
    },
    Block {
        open_bracket: Range,
        close_bracket: Range,
        statements: Vec<ExprPatP>,
    },
    Borrow {
        borrow_token: Range,
        expr: Box<ExprPatP>,
    },
    Explicit {
        explicit_token: Range,
        expr: Box<ExprPatP>,
    },
    Copy {
        copy_token: Range,
        expr: Box<ExprPatP>,
    },
    ConstructData {
        data_constructor: IdentifierP,
        open_brace: Range,
        close_brace: Range,
        fields: ConstructDataFields,
    },
    Impl {
        impl_token: Range,
        body: DefinitionBodyP,
    },
    ImplPattern {
        impl_token: Range,
        open_brace: Range,
        close_brace: Range,
        fields: ConstructDataFields,
    },
    Match {
        match_token: Range,
        expr: Box<ExprPatP>,
        cases: Vec<(ExprPatP, ExprPatP)>,
    },
    Unknown(Range),

}

Represents either an expression or a pattern.

Variants

Variable(IdentifierP)

Tuple Fields

0: IdentifierP

A named variable e.g. x or +.

Constant

Fields

range: Range

value: ConstantValue

A primitive constant such as 14 or false.

String

Fields

range: Range

value: String

A string literal, which will later be converted into a cons-list.

Apply(Box<ExprPatP>, Box<ExprPatP>)

Tuple Fields

0: Box<ExprPatP>

1: Box<ExprPatP>

Apply the left hand side to the right hand side, e.g. a b. More complicated expressions e.g. a b c d can be desugared into ((a b) c) d.

Lambda

Fields

lambda_token: Range

params: Vec<NameP>

expr: Box<ExprPatP>

A lambda abstraction, specifically lambda params -> expr.

Let

Fields

let_token: Range

name: NameP

expr: Box<ExprPatP>

A let statement, specifically let identifier = left_expr;. This kind of statement is only valid as an intermediate statement in a block.

Block

Fields

open_bracket: Range

close_bracket: Range

statements: Vec<ExprPatP>

A block of statements, inside parentheses, separated by newlines, where the final expression in the block is the type of the block as a whole.

Borrow

Fields

borrow_token: Range

expr: Box<ExprPatP>

Borrow some data for a duration less than its full lifetime.

Explicit

Fields

explicit_token: Range

expr: Box<ExprPatP>

Don’t add default impls to this expression.

Copy

Fields

copy_token: Range

expr: Box<ExprPatP>

Copy some data behind a borrow.

ConstructData

Fields

data_constructor: IdentifierP

open_brace: Range

close_brace: Range

fields: ConstructDataFields

The name of a data type or enum variant, followed by brace brackets containing the data structure’s fields.

Impl

Fields

impl_token: Range

body: DefinitionBodyP

An implementation of an aspect. Only used in expressions, not patterns (see ExprPatP::ImplPattern).

ImplPattern

Fields

impl_token: Range

open_brace: Range

close_brace: Range

fields: ConstructDataFields

A pattern destructuring an implementation of an aspect. Only used in patterns, not expressions (see ExprPatP::Impl).

Match

Fields

match_token: Range

expr: Box<ExprPatP>

cases: Vec<(ExprPatP, ExprPatP)>

A list of patterns and their replacements.

A match expression, specifically something of the form match expr { pat -> result, pat -> result, ... }

Unknown(Range)

Tuple Fields

0: Range

An underscore _ representing an unknown. This is valid only in patterns, not normal expressions.